requirements.txt
langchain
langchain-openai
langgraph
python-dotenv
openai
pydantic_v1
pydantic



# App
flask